Update contrib's bootstrap and build process to support QNX.
This mostly involved recognizing a host triplet with 'nto' in it
and then ensuring that an SDP has been sourced into the environment.

Once this is verified, both the HAVE_QNX and HAVE_LINUX flags are set.
QNX and Linux are often close enough that what works on Linux works on
QNX. So to keep from having to touch everything HAVE_LINUX is set. In
cases where something needs to handle different behaviour between Linux
and QNX, it can switch on the HAVE_QNX flag.

Finally QNX's cross compilers (qcc/q++) are set up with the required
environment variables so that they know what target they are compiling
for.

Unfortunately, meson doesn't play nice with the QNX cross compilers. It
doesn't know how to identify them so refuses to build anything. For
meson, call the underlying toolchain directly instead of going through
the qcc/q++ frontends.

This commit does NOT actually make any updates to the individual contrib
libraries. That will be happening in the following commits.