Apply pkg_static to SDL_images's .pc file
For libtool to properly link against SDL_image's archive, all
dependencies have to be public. Apply the pkg_static.sh script against
SDL_image's .pc file so that any private dependencies are turned into
public ones.

To do this, the make had to be run as a two-step. One to build and one
to install. That way the .pc file can be modified in between the two
steps.