Disable contrib/bluray library for QNX builds
The bluray library requires a JDK to build against and I'm not aware of
one currently available for QNX. If one is found later it can be
renabled.

I did start on trying to get bluray to compile and it ran into a
namespace conflict with the various _mutex_* APIs. So this also includes
a patch to rename libbluray's internal _mutex_* APIs to _bd_mutex_*.