Get contrib/ffmpeg building for QNX
ffmpeg's build process needed to be taught how to build with qcc/q++.
Mostly so that it knew the correct flags to pass to the compiler in
order to get the preprocessor to generate the expected dependency files.

In addition, when building for x86_64, the assembler didn't like some of
the hand written assembly code. Turns out there was already a fix for
that in ffmpeg's upstream, so I'm adding it as part of this change.