Get contrib/fribidi building on QNX
fribidi's meson build adds the -ansi flag to the build if it detects
that the compiler supports it. I'm not sure why.

QNX8 no longer supports this option. Normally meson's compiler feature
detection would pick up on this but because the meson build process
in the docker doesn't work with qcc/q++, I have to specify the
underlying compiler directly. That compiler _does_ support "-ansi" so it
gets turned on.

The build then fails when including stdlib.h because it makes use of
"inline" which "-ansi" does not allow.

Remove the usage of "-ansi" when QNX is in play.