Get contrib/gnutls working for QNX builds
gnutls's .pc file was being modified to make private
libraries/dependencies public. That way when VLC creates the eventual
.so file, all the required dependencies are picked up by libtool.

However, this was being done after the tarball was extracted but before
the build was run. The file it was applied to is used during the build
to generate the actual .pc file. That input file didn't match the form
that pkg-static.sh expected so it didn't actually do anything.

Change the build order a bit so that pkg-static.sh is applied to the
actual .pc file, after it has been generated by the build.