Get contrib/gpg-error building with QNX
gpg-error needs to pass the -P flag to the preprocessor via -Wp,-P
instead of directly to qcc, which assigns a different meaning to -P.

Create a local header for QNX8 support that gets copied into gpg-error's
source tree. It is specific to QNX8 though it would probably work fine
for QNX 7.1. If/when support for QNX7.1 is required it can be dealt with
then.