Get contrib/chromaprint working for QNX
Chromaprint, by default, only enables -fpic when it is building the
shared library variant of itself. VLC's contrib builds static libraries
that are then used to build shared libraries. So -fpic isn't used even
though it needs to be.

Add a CMake option to force -fpic to be used. This switch is enabled for
QNX.