Get contrib/srt working on QNX8
Add a QNX target type that can be switched on. It is generally treated
the same as the POSIX type but changes a few places where different
header files need to be included.

In addition, make sure that when static libraries are built, they can be
built with -fpic so that VLC can turn them into shared libraries later.