Fix build when hand-written assembler doesn't use GOT
Several of the contrib libraries (eg: libavcodec) have hand written
assembler in them. While the c/c++ code is compiled with -fPIC, the hand
written assembly generally isn't and doesn't bother trying to access
symbols it wants via the GOT.

This causes a problem when creating the final shared objects as the
(non-GOT) relocation is not supported.

Using SYMBOLIC_LDFLAGS adds -Bsymbolic to the link command line which
in turn makes the linker treat global symbols that it finds as local
and hence don't need the GOT.

This does mean they can't be overriden by the dynamic linker but I
believe that is ok in this instance.