Audio playback locks up on QNX
On QNX, sometimes audio playback would lock up. The easiest way to
trigger this would be to start playback, then stop/restart it a few
times.

The problem was traced to how QNX deals with thread cancellation. It is
a little different than Linux. Once a thread with a cancellation type of
DEFERRED is cancelled, it _will_ be cancelled at the next cancellation
point. Regardless of what the cancellation state of the thread is at the
time the cancellation point is hit.

The upswing is that if the audio_output's worker thread already had a
deferred cancellation request when it set the cancellation state to
DISABLED, it would get cancelled at the next cancellation point, while
it held the mutex lock. Holding this lock would prevent future audio
playback from working.

Work around this by explicitly forcing a cancellation point every time
the cancellation state is disabled. If cancellation is already disabled,
the call correctly does nothing, so it really only has an effect when
called when cancellation is currently enabled.

This solution isn't foolproof. It is possible, though unlikely, that a
cancellation could happen between the call to pthread_testcancel and the
call to pthread_setcancelstate(DISABLED). Unfortunately I couldn't find
another method to trigger a cancellation point that was lightweight and
reliable.