Get contrib/gcrypt working on QNX
When building QNX for an x86_64 target, gcrypt's configure process
thinks that underscores get added to external symbols. This isn't true
and causes some of the custom assembly code to not be found.

Force configure to keep the underscore setting off. This is really only
for x86_64 builds but is safe to apply to aarch64 builds as that also
wants underscores to be off.