Plug VLC's build into QNX recursive build
Provides required directory structure and Makefiles to allow VLC to be
built via QNX's recursive build system.

There is a linux-x86_64 target that does some work that is required to
enable the actual target builds. Mostly downloading all the contrib
packages and compiling a host version of protoc. The former is required
to avoid concurrency conflicts when running with multiple make
processes. The latter because the protoc that comes in modern Ubuntu
installs (like the one in the docker image) is too new.