Fix broken build
Fix a couple of build breaks

(1)
The host protoc build was failing when using a newer docker container.
This is probably because it has a newer version of libstdc++ in it.
Luckily, protobuf already had a fix for it (from 2017!) that I can apply
to make the failure go away.

(2)
libgpg-error was failing to build. It was supposed to look for
lock-obj-pub.nto-qnx8.0.0.h but instead it was looking for
lock-obj-pub.nto-qnx.h.

I'm not sure why this changed, or even if it changed. Perhaps I had a
stale file before that was allowing it to pass. In any case, the two
files are identical at the moment. Just check in a versions under both
names to get things going and (possibly) protect against older systems
wanting the old name.